experts/yWorkflows/component_creator/requirements.txt
ptvsd==4.3.2
pyyaml==6.0
pyyaml-include==1.3
fastapi==0.92.0
python-dotenv==1.0.0
openai==0.27.0
numpy==1.24.2